fix: fixed bounded context structure.

+++ PERUSTARS/PERUSTARS/Shared/Domain/Repositories/IBaseRepository.cs
@@ -7,7 +7,7 @@ namespace PERUSTARS.Shared.Domain.Repositories
     public interface IBaseRepository<TEntity>
     {
         Task AddAsync(TEntity entity);
-        Task<TEntity?> FindByIdAsync(BigInteger id);
+        Task<TEntity?> FindByIdAsync(long id);
         void Update(TEntity entity);
         void Remove(TEntity entity);
         Task<IEnumerable<TEntity>> ListAsync();
